FIX: messages after a philosopher died

+++ srcs/aux/aux.h
@@ -5,6 +5,6 @@
 
 int	ft_atoi(const char *str);
 int	ft_isdigit(int c);
-int	time_diff_us(struct timeval time1, struct timeval time2);
+long long int	time_diff_us(struct timeval time1, struct timeval time2);
 
 #endif
